Remove minijasminenode2 and tests for PhantomJS
minijasminenode2 uses `execSync` which is not
available for versions of Node under `0.12.x`.

Apps using Gulp to build their static assets through
[Travis-CI](https://travis-ci.org/) where Node is
not the primary language cannot specify the Node
version* and so are currently locked to
`v0.10.36`.

*It is possible to install any version of Node as
a pre-build step but as Travis includes PhantomJS
with all its workers this method (which would slow
down build considerably) is unnecessary if we
assume use of it every time.